Move text-utils to pure
