+++ extensions/ql-vscode/src/pure/text-utils.ts
